Add method to display current user location
